fix(travis): do not use system python, instead use pyenv supplied by travis
python2 is officially dead come 2020. Instead of using version 2 or 3 of system python we will
use the travis provided pyenv and use the preinstalled version of python that each travis build
environment spawns with, this allows us to install our own packages without stepping on the system
resulting in a more predictable build.