Overmap Radio Beacons
Not actually used anywhere yet but all is working.
Basically just a small overmap object that can attach to another object and move with it, and give a message when scanned via sensors consoles etc.
Added an item for spawning those, but item itself is not accessible ingame.
Could be used for SOS and crap but lazy.