port progress
Successfully ported over URFC armor from Innie 13, also edited tradeship.dm, which is the default map as an example to see how to get the mod packs to load, also there is several files that can be used to place various sets of clothing code in with the URFC used as a template. Hopefully I make sense here.