Small map tweaks to tradeship
Puts more light switches around so you don't have to cross entire cargo bay to turn lights on.
Makes switch for hallway on deck 1 start as on, cause you need to walk aaaall the way around to switch it and it's awkward.
Puts emergency dispeners around 2nd deck hallway so you ahve a bit of a chance to live during breach.
ecece0