Added decl emote sounds
Basically made a system using decl to load and override via inheritance specie specific emote sounds. The decl itself will load a json file from disk with the sounds for each emotes and optionally genders. Each files is named by specie and is specified in the decl. The idea is to make it simple and quick to use even without a recompile.

It will also be used in our downstream to add more emote sounds.