Fix examine returning that machines not using power were unpowered
Examining those machines would result in the examine text to mention that the machine was not receiving power. Because those are machines that don't need power and don't have tesla link or terminal connection. The issue was that the base machinery code assume things need power unless interact_offline is true. Since none of those machines uses a ui its a bit weird to toggle it to true, but it keeps the message from showing up.