Dislocation immunity set by limb flag now
Dislocation used to use magic numbers to determine if a limb was dislocated, not dislocated or immune to dislocation. While everything else used a limb flag to define this kind of thing. So I made a flag for dislocation instead to bring this stuff closer to the rest of the organ "capabilities"