Limb dislocation fix + improvements (#2337)
* Replaced checks for dislocation

A lot checks for dislocation were comparing against 2 for some reasons, which would never be true.
Replaced all the other checks with a call to is_dislocated()

* Replaced all dislocated check with is_dislocated

* Dislocation immunity set by limb flag now

Dislocation used to use magic numbers to determine if a limb was dislocated, not dislocated or immune to dislocation. While everything else used a limb flag to define this kind of thing. So I made a flag for dislocation instead to bring this stuff closer to the rest of the organ "capabilities"

* un-dislocate limbs on rejuv

* fix typo

* Added pain to dislocate

For some reasons, none of the code would actually cause any sort of pain when dislocating a limb.. Except when un-dislocating. So added some pain + a little bit of weaken, since that hurts a whole lot.

* Added feedback on skillcheck failure

When dislocating a limb, if the grabber doesn't have the required skills it would just say nothing and fail. Even though the current code always warn you you don't have a strong enough grip to dislocate regardless of your skills.
So I added a warning when failing to joint lock or dislocate.

* Change weaken to stun

* Didn't mean to commit this line

* Applied suggested text fixes

* Made dislocation a status flag

* Fix the oopsie I made when rebasing

* Replaced checks for dislocation

A lot checks for dislocation were comparing against 2 for some reasons, which would never be true.
Replaced all the other checks with a call to is_dislocated()

* Replaced all dislocated check with is_dislocated

* un-dislocate limbs on rejuv

* Fix the fix that was fixed before my fix

* The rebase apparently re-duplicated this for some reasons

* Use limb level can_feel_pain proc