Allow items to override their initial health
Allow items to override their initial health

Material used to define what health the item would have in all cases but when the item was invincible. Changed that so the material doesn't replace the non-null health/max_health that was defined.

Moved some of the code in item.dm to better places

* Made a file for item damage stuff since item.dm is pretty crowded
* Removed completely redundant and unused proc /obj/item/proc/get_loc_turf()
* Moved /mob/living/carbon/verb/showoff() to the carbon.dm file since its a better fit.