got rid of the initial reagent list
As lohikar and noelle pointed out using a list for initializing reagents was probably not the best idea because of hidden costs tied to list overrides creating a hidden (init) proc, and the list existing once on each items for their whole existence.
So everything was moved into 2 procs to allow overriding the initial reagents without messing with the initialization of the reagent holder and etc.