Fix ladders from breaking open turfs
Ladders would always call T.ReplaceWithLattice() on init on the turf they're on, even if there is a lattice there already.
But the bad part in this is that ReplaceWithLattice() also changes the turf it's on to the base turf if it isn't already the same type. Which can have pretty bad side effects.

Normally the only thing a ladder should require is to be on an open turf of any type and have a lattice under itself.