Rename syndie cig packs
Gave a temporary initial name to the syndie laced cig packs since they would cause the codex stuff to crash when retrieving info on the pack, because they had the same name as the default cigarette pack.
The name of the default cig pack is then applied back above the initial name.