Added CanClone() proc
Added a better way to tell if we should clone something or not via the CanClone() proc. The listDeeperCopy() will now return a datum by reference if it cannot be cloned, and otherwise just clone it.
Added a check to the base implementation of Clone() to crash when cloning something that cannot be cloned.