Makes door leak damage to components when damaged
Previously doors would soak all damage up to its health and only then start passing it to components.
This made whole thing kinda meaningless because by the time you start damaging copmonents the door is gone, you can just crowbar it open

Now if door is at 75% health or if it's 10 damage in single hit, some damage can leak through.
It's scaled to how broken door is, so you're not going to be crunching locks with first few lucky hits.
Secure airlocks override that to keep older behaviour because that's more secure