Fix overmap having dense turfs
The overmap was within the bottom left corner of the transition edge, and would have its turfs turned solid because of a slight oversight on my part. Corrected the problem by just not making turf inside the corners solid by default. You have to specify a border filler turf to do it now.
* Fixed overmap spawning 2 instances of the base area.
* Added a dark filler turf to hide the area outside the overmap.