Fix template load bound checking
The code would just add the template's width and height to the turf picked as center without subtracting one from the width and height of the template. Which would cause the size check to fail in the cases where the template was the same size as the world size. Since the expected top right corner turf would end up outside the world.
* Added some helper macro for world size stuff to keep things a bit more coherent.