Fixed lockable safes interactions
Since the interaction on the base class to open the lockable ui required to hold the item in a hand, you couldn't open the lockable ui. Made it so, when impossible to pick up, clicking the safe opens the ui.
Ideally, the safe should just be made into a structure, since it essentially works like one, and also so we don't have to mess with the item pickup code.